requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.1
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
pytesseract==0.3.10
Pillow==10.0.1
python-jose==3.3.0
passlib==1.7.4 